- Add human readable file size logic - Remove hard coded MB format space logice

+++ document-service/src/main/java/com/filecloud/documentservice/model/dto/SpaceInfoResponseDto.java
@@ -4,13 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class SpaceInfoResponseDto {
 
-    private double spaceLimit;
-    private double usedSpace;
-    private double remainingSpace;
+	private long spaceLimit;
+	private long usedSpace;
+	private long remainingSpace;
 
 }
